Majorly expand the word list
Switches to a manually assembled word list from SCOWL's huge set of English words.
The new list has 3 times as many words as the previous one did, so there should hopefully be less conspicuously missing words in the data set.
It is worth noting that this may also introduce some weird words that probably shouldn't count.
That being said, the previous word list did as well so it seems like a reasonable tradeoff.